Rename benchmarks/tasks/qwen3_asr.py to asr_transcription.py and fix some nit
